server/requirements.txt
fastapi
uvicorn[standard]
requests
gunicorn
python-dotenv
